requirements.txt
asyncio>=3.4.3
google-genai>=1.15.0
mcp>=1.9.0